Create about.html
Added an "about" page, based on the content from the index.html page, in order to have a page to link to from the navigation menu.   I am creating this file directly in the repository as a temporary measure until I can align the email account on the git terminal with the one for this repository.